backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiohttp==3.9.1
aiofiles==23.2.1
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
xhtml2pdf==0.2.11
pydantic==2.5.2
websockets==12.0
scapy==2.5.0
python-whois==0.9.4
email-validator==2.1.0
